% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisiontreeutils.R
\name{adjustCP}
\alias{adjustCP}
\title{adjusts config based on results if config was initially "Auto"}
\usage{
adjustCP(config, the_model)
}
\arguments{
\item{config}{list of config options}

\item{the_model}{model object}
}
\value{
model obj after adjusting complexity parameter
}
\description{
adjusts config based on results if config was initially "Auto"
}

