% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisiontreeutils.R
\name{adjustCP.rxDTree}
\alias{adjustCP.rxDTree}
\title{Adjusts config based on results if config was initially "Auto"}
\usage{
\method{adjustCP}{rxDTree}(model, config)
}
\arguments{
\item{model}{model object}

\item{config}{list of config options}
}
\description{
Adjusts config based on results if config was initially "Auto"
}

