% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admission}
\alias{admission}
\title{Admission results to graduate school}
\format{A data frame with 400 rows and 4 variables:
\itemize{
  \item admit: if the student is admitted (1: admitted; 0: not admitted)
  \item gre: GRE score.
  \item gpa: undergraduate GPA
  \item rank: class rank
}}
\usage{
admission
}
\description{
A dataset containing if a student is admitted to graduate school.
The data is obtained from: http://www.ats.ucla.edu/stat/data/binary.csv
The variables are as follows:
}
\keyword{datasets}

