% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alteryxreport.R
\name{Alteryx.ReportAnova}
\alias{Alteryx.ReportAnova}
\title{Create a data frame with model object summary}
\usage{
Alteryx.ReportAnova(model.obj)
}
\arguments{
\item{model.obj}{model object whose summary output is put into a data frame}
}
\description{
The function Alteryx.ReportAnova creates a data frame of key-value pairs for
the purpose of assisting Alteryx's reporting tools to report the results of a
type II ANOVA for a LM or GLM model object. The last key- value indicates
whether the report is for a lm or glm model which is relevant for labeling
columns in the output created by Alteryx.
}
\seealso{
Other Alteryx.Report: \code{\link{Alteryx.ReportGLM}},
  \code{\link{Alteryx.ReportLM}},
  \code{\link{createReportGLMNET}}
}
\author{
Dan Putler
}
