% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alteryxreport.R
\name{Alteryx.ReportGLM}
\alias{Alteryx.ReportGLM}
\title{Create a data frame with glm model object summary}
\usage{
Alteryx.ReportGLM(glm.obj)
}
\arguments{
\item{glm.obj}{glm model whose summary output is put into a data frame}
}
\description{
The function Alteryx.ReportGLM creates a data frame of an glm model's summary
output that can more easily be handled by Alteryx's reporting tools. The
function returns a data frame that is a set of key-value pairs for the
components of the summary report as well as the model's coefficients in a
JSON formatted character string. The coefficient JSON string is really for
future option value
}
\author{
Dan Putler
}
\seealso{
Other Alteryx.Report: \code{\link{Alteryx.ReportAnova}},
  \code{\link{Alteryx.ReportLM}},
  \code{\link{createReportGLMNET}}
}

