% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alteryxreport.R
\name{Alteryx.ReportLM}
\alias{Alteryx.ReportLM}
\title{Create a data frame with lm model object summary}
\usage{
Alteryx.ReportLM(lm.obj)
}
\arguments{
\item{lm.obj}{lm model object whose summary output is put into a data frame}
}
\description{
The function Alteryx.ReportLM creates a data frame of an lm model's summary
output that can more easily be handled by Alteryx's reporting tools. The
function returns a data frame that is a set of key-value pairs for the
components of the summary report as well as the model's coefficients in a
JSON formatted character string.
}
\author{
Dan Putler
}
\seealso{
Other Alteryx.Report: \code{\link{Alteryx.ReportAnova}},
  \code{\link{Alteryx.ReportGLM}}
}

