% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testutils.R
\name{Alteryx_assert.default}
\alias{Alteryx_assert.default}
\title{analog of assert_that for use with Alteryx,
defaults to error messaging}
\usage{
\method{Alteryx_assert}{default}(exp, msg, ...)
}
\arguments{
\item{exp}{expression to evaluate}

\item{msg}{message to pass to user}

\item{...}{additional arguments}
}
\description{
analog of assert_that for use with Alteryx,
defaults to error messaging
}

