% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testutils.R
\name{Alteryx_assert}
\alias{Alteryx_assert}
\title{analog of assert_that for use with Alteryx
Generic S3 class on message
message can have class 'error' or 'warning'}
\usage{
Alteryx_assert(exp, msg, ...)
}
\arguments{
\item{exp}{expression to evaluate}

\item{msg}{message to pass to user}

\item{...}{additional arguments}
}
\description{
analog of assert_that for use with Alteryx
Generic S3 class on message
message can have class 'error' or 'warning'
}
