% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testutils.R
\name{Alteryx_assert.warning}
\alias{Alteryx_assert.warning}
\title{analog of assert_that for use with Alteryx warnings}
\usage{
\method{Alteryx_assert}{warning}(exp, msg, ...)
}
\arguments{
\item{exp}{expression to evaluate}

\item{msg}{message to pass to user}

\item{...}{additional arguments}
}
\description{
analog of assert_that for use with Alteryx warnings
}
