% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{AlteryxGraph2}
\alias{AlteryxGraph2}
\title{Plot output of an R expression}
\usage{
AlteryxGraph2(nOutput, expr, config = NULL, plotPrefix = NULL)
}
\arguments{
\item{expr}{code generating a plot}

\item{config}{list containing plot configuration}

\item{plotPrefix}{prefix to use for the plot}
}
\description{
Plot output of an R expression
}

