% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{AlteryxGraph3}
\alias{AlteryxGraph3}
\title{Plot output of an R expression}
\usage{
AlteryxGraph3(expr, nOutput, config = NULL, plotPrefix = NULL)
}
\arguments{
\item{expr}{code generating a plot}

\item{nOutput}{output stream to plot the graph to}

\item{config}{list containing plot configuration}

\item{plotPrefix}{prefix to use for the plot}
}
\description{
Plot output of an R expression
}

