% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{AlteryxMessage2}
\alias{AlteryxMessage2}
\title{Message}
\usage{
AlteryxMessage2(msg, ...)
}
\arguments{
\item{msg}{message}

\item{...}{extra arguments to pass on to AlteryxMessage}
}
\description{
Message
}

