% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alteryxreport.R
\name{AlteryxReportRx}
\alias{AlteryxReportRx}
\title{Create report for rx model objects.}
\usage{
AlteryxReportRx(rx.obj, null.deviance = NULL)
}
\arguments{
\item{rx.obj}{model object of class rxLinMod, rxLogit or rxGlm}

\item{null.deviance}{null deviance}
}
\description{
Create report for rx model objects.
}

