% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactiveutils.R
\name{badDash}
\alias{badDash}
\title{Create a dashboard outputting a single message}
\usage{
badDash(messageIn, titleIn = "Error", colorIn = "aqua", widthIn = 12)
}
\arguments{
\item{messageIn}{message to output}

\item{titleIn}{title of dash}

\item{colorIn}{color in dash}

\item{widthIn}{width}
}
\value{
dashboard object for rendering
}
\description{
Create a dashboard outputting a single message
}
\author{
Todd Morley
}

