% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonutils.R
\name{bootCH}
\alias{bootCH}
\title{Bootstrap Replicates of the Calinski-Harabas index for Cluster Validation}
\usage{
bootCH(xdat, k_vals, clstr1, clstr2, cntrs1, cntrs2, method = c("kmn", "kmd",
  "neuralgas"))
}
\arguments{
\item{xdat}{A numeric matrix of the data to be clustered.}

\item{k_vals}{An integer vector giving the set of clustering solutions to be
examined.}

\item{clstr1}{The cluster assignments from a bootFlexclust object for one
side of the Rand index paired comparisons.}

\item{clstr2}{The cluster assignments from a bootFlexclust object for the
other side of the Rand index paired comparisons}

\item{cntrs1}{The cluster centers from a bootFlexclust object for one side of
the bootFlexclust Rand index paired comparisons.}

\item{cntrs2}{The cluster centers from a bootFlexclust object for the other
side of the bootFlexclust Rand index paired comparisons.}

\item{method}{The clustering method, one of "kmn" (K-Means), "kmd"
(K-Medians), and "neuralgas" (neural gas).}
}
\description{
This function calculates the Calinski-Harabas index for all solutions
contained in a bootFlexclust object. It does it for each Rand test paired
comparison, so 100 bootstrap replicates of the Rand index will result in 200
Calinski-Harabase index values. This code borrows from the function index.G1
of Marek Walesiak and Andrzej Dudek's clusterSim package in terms of
implementing the sum of squares components
}
\author{
Dan Putler
}

