% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtree_cross_validation.R
\name{checkFactorVars}
\alias{checkFactorVars}
\title{For each factor variable, check to see if all levels are present in each fold.
If not, warn the user.}
\usage{
checkFactorVars(data, folds, config)
}
\arguments{
\item{data}{a data.frame with the data used to generate the models}

\item{folds}{a list of record id's in each fold in each trial (list of lists)}

\item{config}{a list of configuration information}
}
\description{
For each factor variable, check to see if all levels are present in each fold.
If not, warn the user.
}

