% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkInstalls}
\alias{checkInstalls}
\title{Utility function to check to see if the necessary packages are
installed and install them if they're not.}
\usage{
checkInstalls(packages)
}
\arguments{
\item{packages}{vector of package names to check and install.}
}
\description{
Utility function to check to see if the necessary packages are
installed and install them if they're not.
}
\author{
Bridget Toomey, Dan Putler
}
