% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defensiveutils.R
\name{checkLowN}
\alias{checkLowN}
\title{Check if data has enough rows}
\usage{
checkLowN(data, threshold = 25, mult = 1,
  msg = paste0("The incoming data may not have ",
  "enough data to generate a model succesfully."))
}
\arguments{
\item{data}{dataframe}

\item{threshold}{minimum number of records needed to not give warning}

\item{mult}{multiplier portion of data used for model training}

\item{msg}{output message (warning)}
}
\value{
no return - only throws warning
}
\description{
Check if data has enough rows
}

