% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defensiveutils.R
\name{checkMissing.omit}
\alias{checkMissing.omit}
\title{Check if data contains missing values, omit those records}
\usage{
checkMissing.omit(data)
}
\arguments{
\item{data}{dataframe to check}
}
\value{
data with missing records excluded
}
\description{
Check if data contains missing values, omit those records
}

