% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisiontreeutils.R
\name{checkValidConfig}
\alias{checkValidConfig}
\title{Error checking pre-model
Does not return anything - just throws errror}
\usage{
checkValidConfig(config, the.data, names, is_XDF)
}
\arguments{
\item{config}{list of config options}

\item{the.data}{incoming data}

\item{names}{list of x, y, w names for data}

\item{is_XDF}{whether data is XDF}
}
\description{
Error checking pre-model
Does not return anything - just throws errror
}

