% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisiontreeutils.R
\name{checkValidConfigrxDTree}
\alias{checkValidConfigrxDTree}
\title{Error checking pre-model
Does not return anything - just throws errror}
\usage{
checkValidConfigrxDTree(config, the.data, names)
}
\arguments{
\item{config}{list of config options}

\item{the.data}{incoming data}

\item{names}{list of x, y, w names for data}
}
\description{
Error checking pre-model
Does not return anything - just throws errror
}
