% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_DT.R
\name{classification_mismatches}
\alias{classification_mismatches}
\title{The classification_mismatches function produces a sparse confusion
matrix listing misclassifications in descending order of frequency,
and limiting its output to some total cumulative percentage of
misclassifications (in case there's a long tail of low-probability
misclassifications).  This lets the user try to alter the model
in ways calculated to achieve improvement around specific, frequently
occurring cases of misclassification.}
\usage{
classification_mismatches(actual_values, fitted_values, digits = 3,
  cutoff_cumulative_percent = 80)
}
\arguments{
\item{actual_values}{vector of expected values}

\item{fitted_values}{vector of fitted values}

\item{digits}{number of digits past the decimal to round to}

\item{cutoff_cumulative_percent}{cumulative percentage at which to stop outputting}
}
\value{
dataframe with at least one row truncated at the cumulative percent
}
\description{
The classification_mismatches function produces a sparse confusion
matrix listing misclassifications in descending order of frequency,
and limiting its output to some total cumulative percentage of
misclassifications (in case there's a long tail of low-probability
misclassifications).  This lets the user try to alter the model
in ways calculated to achieve improvement around specific, frequently
occurring cases of misclassification.
}
\author{
Todd Morley
}

