% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtree_cross_validation.R
\name{computeBinaryMetrics}
\alias{computeBinaryMetrics}
\title{Get the necessary measures in the binary classification case}
\usage{
computeBinaryMetrics(pred_prob, actual, threshold)
}
\arguments{
\item{pred_prob}{vector of predicted probabilities}

\item{actual}{vector of actual results}
}
\value{
a data.frame with results
}
\description{
Get the necessary measures in the binary classification case
}

