% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisiontreeutils.R
\name{convertDTParamsToArgs}
\alias{convertDTParamsToArgs}
\title{Map parameter names to function argument names}
\usage{
convertDTParamsToArgs(params)
}
\arguments{
\item{params}{list of decision tree params}
}
\value{
list with named parameters  for function
}
\description{
Map parameter names to function argument names
}

