% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisiontreeutils.R
\name{convertDTParamsToArgs.rpart}
\alias{convertDTParamsToArgs.rpart}
\title{Map parameter names to function arg names for rpart}
\usage{
\method{convertDTParamsToArgs}{rpart}(params)
}
\arguments{
\item{params}{list of decision tree params}
}
\value{
list with named parameters for rpart
}
\description{
Map parameter names to function arg names for rpart
}

