% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisiontreeutils.R
\name{convertDTParamsToArgs.rxDTree}
\alias{convertDTParamsToArgs.rxDTree}
\title{Map parameter names to function arg names for rxDTree}
\usage{
\method{convertDTParamsToArgs}{rxDTree}(params)
}
\arguments{
\item{params}{list of decision tree params}
}
\value{
list with named parameters for rxDTree
}
\description{
Map parameter names to function arg names for rxDTree
}

