% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisiontreeutils.R
\name{convertDTParamsToArgsC5.0}
\alias{convertDTParamsToArgsC5.0}
\title{Map parameter names to function arg names for C5.0}
\usage{
convertDTParamsToArgsC5.0(params)
}
\arguments{
\item{params}{list of decision tree params}
}
\value{
list with named parameters for C5.0
}
\description{
Map parameter names to function arg names for C5.0
}
