% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisiontreeutils.R
\name{convertDTParamsToArgsrpart}
\alias{convertDTParamsToArgsrpart}
\title{Map parameter names to function arg names for rpart}
\usage{
convertDTParamsToArgsrpart(params)
}
\arguments{
\item{params}{list of decision tree params}
}
\value{
list with named parameters for rpart
}
\description{
Map parameter names to function arg names for rpart
}
