% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisiontreeutils.R
\name{convertDTParamsToArgsrxDTree}
\alias{convertDTParamsToArgsrxDTree}
\title{Map parameter names to function arg names for rxDTree}
\usage{
convertDTParamsToArgsrxDTree(params)
}
\arguments{
\item{params}{list of decision tree params}
}
\value{
list with named parameters for rxDTree
}
\description{
Map parameter names to function arg names for rxDTree
}

