% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisiontreeutils.R
\name{createDashboardDT.default}
\alias{createDashboardDT.default}
\title{Create Interactive Dashboard for default model}
\usage{
\method{createDashboardDT}{default}(model)
}
\arguments{
\item{model}{model object}
}
\description{
Create Interactive Dashboard for default model
}
