% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisiontreeutils.R
\name{createDashboardDT}
\alias{createDashboardDT}
\title{Generic S3 Class
Create Interactive Dashboard}
\usage{
createDashboardDT(model)
}
\arguments{
\item{model}{model object}
}
\description{
Generic S3 Class
Create Interactive Dashboard
}

