% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisiontreeutils.R
\name{createDashboardDT.rpart}
\alias{createDashboardDT.rpart}
\title{Create Interactive Dashboard for rpart model}
\usage{
\method{createDashboardDT}{rpart}(model)
}
\arguments{
\item{model}{model object}
}
\description{
Create Interactive Dashboard for rpart model
}
