% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisiontreeutils.R
\name{createDTParams}
\alias{createDTParams}
\title{Creation of components for model object evaluation}
\usage{
createDTParams(config, names)
}
\arguments{
\item{config}{list of config options}

\item{names}{list of variable names (x, y and w)}
}
\value{
list with components needed to create model
}
\description{
Creation of components for model object evaluation
}
