% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVutils.R
\name{createFolds}
\alias{createFolds}
\title{Create the list of cross-validation folds}
\usage{
createFolds(data, config, set_seed = TRUE, seed = NULL)
}
\arguments{
\item{data}{the data.frame used to create the models}

\item{config}{a list of configuration information}

\item{set_seed}{boolean of whether to (re)set seed}

\item{seed}{integer value of random seed}
}
\value{
list of record ID's. Each element is the record ID's of the folds for a given trial.
}
\description{
Create the list of cross-validation folds
}

