% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearregutils.R
\name{createPlotOutputsLinearOSR}
\alias{createPlotOutputsLinearOSR}
\title{Create Plots}
\usage{
createPlotOutputsLinearOSR(the.model)
}
\description{
Prepare the basic regression diagnostic plots if it is requested
and their isn't the combination of singularities and the use of
sampling weights.
}

