% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticregutils.R
\name{createPlotOutputsLogisticOSR}
\alias{createPlotOutputsLogisticOSR}
\title{Create Plots}
\usage{
createPlotOutputsLogisticOSR(the.model, singular, config)
}
\description{
Prepare the basic regression diagnostic plots if it is requested
and their isn't the combination of singularities and the use of
sampling weights.
}

