% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisiontreeutils.R
\name{createPrunePlotDT.C5.0}
\alias{createPrunePlotDT.C5.0}
\title{Create Prune Plot}
\usage{
\method{createPrunePlotDT}{C5.0}(model)
}
\arguments{
\item{model}{model object}
}
\description{
Create Prune Plot
}

