% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisiontreeutils.R
\name{createPrunePlotDT.rxDTree}
\alias{createPrunePlotDT.rxDTree}
\title{Create Prune Plot}
\usage{
\method{createPrunePlotDT}{rxDTree}(model)
}
\arguments{
\item{model}{model object}
}
\description{
Create Prune Plot
}
