% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisiontreeutils.R
\name{createReportDT}
\alias{createReportDT}
\title{Generic S3 class
Get data for static report (grp|out pipes)}
\usage{
createReportDT(model, config = NULL, names = NULL, xdf_path = NULL)
}
\arguments{
\item{model}{model object}

\item{config}{list of config options}

\item{names}{names of variables (x, y and w)}

\item{xdf_path}{string of xdf file location}
}
\value{
dataframe of piped results
}
\description{
Generic S3 class
Get data for static report (grp|out pipes)
}
