% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearregutils.R
\name{createReportGLMNET}
\alias{createReportGLMNET}
\title{Create a data frame with elnet/cv.glmnet containing an elnet model object summary}
\usage{
createReportGLMNET(glmnet_obj)
}
\arguments{
\item{glmnet_obj}{glmnet or cv.glmnet model object whose non-zero coefficients are
put into a data frame}
}
\description{
The function createReportGLMNET creates a data frame of an elnet/cv.glmnet model's summary
output that can more easily be handled by Alteryx's reporting tools. The
function returns a data frame containing the model's coeffcients.
}
\author{
Bridget Toomey
}
\seealso{
Other Alteryx.Report: \code{\link{Alteryx.ReportAnova}},
  \code{\link{Alteryx.ReportGLM}},
  \code{\link{Alteryx.ReportLM}}
}

