% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticregutils.R
\name{createReportLogisticOSR}
\alias{createReportLogisticOSR}
\title{Create Reports}
\usage{
createReportLogisticOSR(the.model, config, model_type)
}
\arguments{
\item{the.model}{model object}

\item{config}{tool configuration}

\item{model_type}{class of model}
}
\description{
If the ANOVA table is requested then create it and add its results to the
key-value table. Its creation will be surpressed if the car package isn't
present, if their were singularities in estimation, or if the input is an
XDF file.
}

