% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticregutils.R
\name{createReportLogisticXDF}
\alias{createReportLogisticXDF}
\title{Create report for logistic regression run on XDF}
\usage{
createReportLogisticXDF(the.model, config, null.model)
}
\arguments{
\item{the.model}{model object}

\item{config}{tool configuration}

\item{null.model}{null model created}
}
\description{
Create report for logistic regression run on XDF
}

