% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisiontreeutils.R
\name{createTreePlotDT.default}
\alias{createTreePlotDT.default}
\title{Create Tree Plot for C5.0 model}
\usage{
\method{createTreePlotDT}{default}(model, config, inputs)
}
\arguments{
\item{model}{model object}

\item{config}{configuration object}

\item{inputs}{input list}
}
\value{
graphs
}
\description{
Create Tree Plot for C5.0 model
}
