% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularizationutils.R
\name{df2NumericMatrix}
\alias{df2NumericMatrix}
\title{Convert data frame into a numeric matrix, filtering out non-numeric columns
and warning the user (by displaying a message passed in as an argument) when
such filtering occurs.}
\usage{
df2NumericMatrix(x, filtering_message)
}
\arguments{
\item{x}{data frame to coerce to a numeric matrix}
}
\description{
Convert data frame into a numeric matrix, filtering out non-numeric columns
and warning the user (by displaying a message passed in as an argument) when
such filtering occurs.
}

