% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createmodelutils.R
\name{doFunction}
\alias{doFunction}
\title{Execute function on parameters
To-Do: re-specify call to be appropriate}
\usage{
doFunction(f, param_list)
}
\arguments{
\item{f}{function string}

\item{param_list}{list of parameters to pass to f}
}
\value{
model object
}
\description{
Execute function on parameters
To-Do: re-specify call to be appropriate
}

