% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactiveutils.R
\name{errorDash}
\alias{errorDash}
\title{Writes default error dashboard}
\usage{
errorDash(titleIn = "Error", colorIn = "aqua", widthIn = 12)
}
\arguments{
\item{titleIn}{title of dash}

\item{colorIn}{color in dash}

\item{widthIn}{width}
}
\value{
dashboard object for rendering
}
\description{
Writes default error dashboard
}

