% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactiveutils.R
\name{getBinaryConfusionMatrix}
\alias{getBinaryConfusionMatrix}
\title{Helpful wrapper around fitted and actual values for generating confusion matrix}
\usage{
getBinaryConfusionMatrix(fitted_values, actual_values)
}
\arguments{
\item{fitted_values}{fitted values}

\item{actual_values}{actual values}
}
\value{
confusion matrix
}
\description{
Helpful wrapper around fitted and actual values for generating confusion matrix
}

