% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisiontreeutils.R
\name{getDTGraphCalls}
\alias{getDTGraphCalls}
\title{get graphing calls}
\usage{
getDTGraphCalls(config, the_model, is_XDF)
}
\arguments{
\item{config}{list of config options}

\item{the_model}{model object}

\item{is_XDF}{boolean of whether model is XDF}
}
\value{
params to call AlteryxGraph on
}
\description{
get graphing calls
}

