% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisiontreeutils.R
\name{getDTPipes}
\alias{getDTPipes}
\title{get grp|out pipes for outputting static report}
\usage{
getDTPipes(config, model, is_XDF)
}
\arguments{
\item{config}{list of config options}

\item{model}{model object}

\item{is_XDF}{boolean of whether model is XDF}
}
\value{
dataframe of piped results
}
\description{
get grp|out pipes for outputting static report
}

