% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisiontreeutils.R
\name{getDTViz}
\alias{getDTViz}
\title{get component for interactive viz}
\usage{
getDTViz(model, is_XDF)
}
\arguments{
\item{model}{model object}

\item{is_XDF}{boolean of whether model is XDF}
}
\description{
get component for interactive viz
}

