% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularizationutils.R
\name{getFamily}
\alias{getFamily}
\title{Figure out which family to use}
\usage{
getFamily(inputs, config)
}
\arguments{
\item{inputs}{input data streams passed to tool}

\item{config}{configuration passed to tool}
}
\value{
string family
}
\description{
Figure out which family to use
}

