% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVutils.R
\name{getMeasuresClassification}
\alias{getMeasuresClassification}
\title{Get the necessary measures in the classification case}
\usage{
getMeasuresClassification(outData, extras)
}
\arguments{
\item{outData}{scored data used to obtain the measures}

\item{extras}{list of miscellaneous information}
}
\value{
outvec a vector of results
}
\description{
Get the necessary measures in the classification case
}

