% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVutils.R
\name{getPosClass}
\alias{getPosClass}
\title{In the 2-class classification case, get the positive class. Otherwise, do nothing.}
\usage{
getPosClass(yVar, order)
}
\arguments{
\item{yVar}{to set Positive class for}

\item{order}{string - choices are "alpha" and "common".
For "alpha", use first by alphabetical order.
For "common", use less common class.}
}
\value{
string - name of positive class
}
\description{
In the 2-class classification case, get the positive class. Otherwise, do nothing.
}

