% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVutils.R
\name{getPositiveClass}
\alias{getPositiveClass}
\title{Get the postive class for two-class classification, choosing the
positive class to the less-common class when all else fails.}
\usage{
getPositiveClass(target_levels)
}
\arguments{
\item{target_levels}{a vector of strings with the target variable's levels}
}
\value{
string - name of positive class
}
\description{
Get the postive class for two-class classification, choosing the
positive class to the less-common class when all else fails.
}

