% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisiontreeutils.R
\name{getReportObjectDT}
\alias{getReportObjectDT}
\title{Get common report objects}
\usage{
getReportObjectDT(model, out)
}
\arguments{
\item{model}{model object}

\item{out}{results from printcp}
}
\value{
list of piped results
}
\description{
Get common report objects
}

