% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogRegCV.R
\name{getResultsCrossValidationLogReg}
\alias{getResultsCrossValidationLogReg}
\title{get results from running cross validation for logistic regression model}
\usage{
getResultsCrossValidationLogReg(inputs, config)
}
\arguments{
\item{inputs}{inputs list with data and models}

\item{config}{list of configuration options}
}
\value{
list of results for outputting
}
\description{
get results from running cross validation for logistic regression model
}

