% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{getWHRParams}
\alias{getWHRParams}
\title{Get list of arguments to pass to whrParams}
\usage{
getWHRParams(config, plotPrefix = NULL)
}
\arguments{
\item{config}{list containing plot related configuration elements}

\item{plotPrefix}{prefix to use for plot, defaults to NULL}
}
\description{
Get list of arguments to pass to whrParams
}
