% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xdfutils.R
\name{getXdfProperties}
\alias{getXdfProperties}
\title{extract xdf properties from incoming xdf data}
\usage{
getXdfProperties(input_name)
}
\arguments{
\item{input_name}{name of incoming connection}
}
\value{
list with boolean is_XDF and string xdf_path representing
 whether data is XDF and the path of the data
}
\description{
extract xdf properties from incoming xdf data
}

