% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreutils.R
\name{getXVars}
\alias{getXVars}
\title{Extract predictor variables from an R model call object}
\usage{
getXVars(x)
}
\arguments{
\item{x}{model object}
}
\description{
Extract predictor variables from an R model call object
}

