% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreutils.R
\name{getYlevels}
\alias{getYlevels}
\title{Get Y levels from model object}
\usage{
getYlevels(x, ...)
}
\arguments{
\item{x}{model object.}

\item{...}{other parameters to pass to the function.}
}
\description{
Get Y levels from model object
}
