% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getYVar.R
\name{getYVar}
\alias{getYVar}
\alias{getYVar.cv.glmnet}
\alias{getYVar.default}
\alias{getYVar.elnet}
\alias{getYVar.naiveBayes}
\title{getYVar function}
\usage{
getYVar(mod)

\method{getYVar}{default}(mod)

\method{getYVar}{elnet}(mod)

\method{getYVar}{cv.glmnet}(mod)

\method{getYVar}{naiveBayes}(mod)
}
\arguments{
\item{mod}{model object}
}
\description{
getYVar function
}
\author{
Ramnath Vaidyanathan, Bridget Toomey
}

