% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVutils.R
\name{getYvars}
\alias{getYvars}
\title{Check if response variable is the same in the pre-built model(s) and the input data.
If so, output this variable.}
\usage{
getYvars(data, models)
}
\arguments{
\item{data}{incoming data}

\item{model(s)}{to extract Y vars from}
}
\value{
y variable
}
\description{
Check if response variable is the same in the pre-built model(s) and the input data.
If so, output this variable.
}

