% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonutils.R
\name{graphWHR2}
\alias{graphWHR2}
\title{Computes width, height and dpi for a graph output}
\usage{
graphWHR2(inches = TRUE, in.w = 5.5, in.h = 5.75, cm.w = NULL,
  cm.h = NULL, graph.resolution = c("1x", "2x", "3x"), print.high = FALSE)
}
\arguments{
\item{inches}{boolean indicating whether to measure dimensions in inches}

\item{in.w}{the width of the plot to produce in inches.}

\item{in.h}{the height of the plot to produce in inches.}

\item{cm.w}{the width of the plot to produce in cm.}

\item{cm.h}{the height of the plot to produce in cm.}

\item{graph.resolution}{the resolution of the plot to produce as a multiple
of 96dpi}

\item{print.high}{whether to turn 3x into 6x (high, 576 dpi) resolution}
}
\description{
This takes information about the desired graph size (in
inches or centemeters), the desired output resolution of 1x, 2x, 3x from a
base of 96 dpi, and and (in the case of 3x) whether high (576 dpi) resolution
should be used to determine the appropriate width, height, and res values to
provide the AlteryxGraph device.
}
\author{
Dan Putler
}

