% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_DT.R
\name{interactive_dt}
\alias{interactive_dt}
\title{Dashboard generation for decision tree models}
\usage{
interactive_dt(config, data, model)
}
\arguments{
\item{config}{list of configuration elements}

\item{data}{dataframe}

\item{model}{model object - should be one of
rpart
C5.0}
}
\value{
dashboard object for rendering
}
\description{
Dashboard generation for decision tree models
}
\author{
Todd Morley

Todd Morley
}
