% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactiveLMReport.R
\name{interactive_lm_report}
\alias{interactive_lm_report}
\title{linear-model interactive-report function
This function generates the interactive report for the Linear Model tool.}
\usage{
interactive_lm_report(config, data, model, cv_metrics = NULL)
}
\arguments{
\item{config}{config}

\item{data}{data (breaking the usual mold a little)}

\item{model}{model}

\item{cv_metrics}{vector of CV model-performance metrics:
R2, adj R2, avg MAE, avg MAPE, avg MSE, avg RMSE}
}
\description{
linear-model interactive-report function
This function generates the interactive report for the Linear Model tool.
}
\author{
Todd Morley, Dylan Blanchard
}

