% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_lr.R
\name{interactive_lr}
\alias{interactive_lr}
\title{Dashboard generation for logistic regression models}
\usage{
interactive_lr(config, data, model)
}
\arguments{
\item{config}{list of configuration elements}

\item{data}{dataframe}

\item{model}{model object - should be one of
glm - with binomial family
glmnet
cv.glmnet}
}
\value{
dashboard object for rendering
}
\description{
Dashboard generation for logistic regression models
}
\author{
Todd Morley
}

