% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testutils.R
\name{is.boundedReal}
\alias{is.boundedReal}
\title{checks whether real value is between min and max}
\usage{
is.boundedReal(val, min = -Inf, max = Inf, closed = TRUE)
}
\arguments{
\item{val}{value to check}

\item{min}{lower bound}

\item{max}{upper bound}

\item{closed}{logical or length-2 logical vector whether range is closed}
}
\description{
checks whether real value is between min and max
}
