% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testutils.R
\name{is.integerValue}
\alias{is.integerValue}
\title{check whether some number is an integerValue}
\usage{
is.integerValue(x, precision = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{numeric to attempt to coerce to whole number}

\item{precision}{precision to compare to nearest whole number}
}
\description{
check whether some number is an integerValue
}

