% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createmodelutils.R
\name{makeFormula}
\alias{makeFormula}
\title{Create formula string}
\usage{
makeFormula(x_vars, y_var)
}
\arguments{
\item{x_vars}{vector of strings of x variable field names}

\item{y_var}{string of y variable field name}
}
\description{
Create formula string
}

