% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoinput.R
\name{makeInput}
\alias{makeInput}
\title{Make an input string to be inserted at the top of R code}
\usage{
makeInput(template, input)
}
\arguments{
\item{template}{template}

\item{input}{input}
}
\description{
Make an input string to be inserted at the top of R code
}
