% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreutils.R
\name{matchLevels}
\alias{matchLevels}
\title{Match levels of factors in original data with new data}
\usage{
matchLevels(nd, ol)
}
\arguments{
\item{nd}{new data}

\item{ol}{old factor levels}
}
\description{
Coerce the levels in new data factors to exactly match the levels
of factors in the original data, and is needed for Revo ScaleR models
}

