% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonutils.R
\name{matrixPipeDelim}
\alias{matrixPipeDelim}
\title{Convert a matrix into a pipe delimited set of values}
\usage{
matrixPipeDelim(a.matrix, round.level = 5e-07)
}
\arguments{
\item{a.matrix}{matrix to convert}

\item{round.level}{rounding level to use}
}
\description{
The matrixPipeDelim function takes a matrix and converts it into a character
vector in which each element is a pipe delimited set of values for a row in
the original matrix
}
\author{
Dan Putler
}
