% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createmodelutils.R
\name{modelOut}
\alias{modelOut}
\title{Prepare model for output}
\usage{
modelOut(model_name, model_obj)
}
\arguments{
\item{model_name}{string - name of model to be output}

\item{model_obj}{model object to output}
}
\value{
key value pair list with name and model
}
\description{
Prepare model for output
}

