% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs.R
\name{mrsCorrectLevels}
\alias{mrsCorrectLevels}
\title{Correct MRS levels.}
\usage{
mrsCorrectLevels(e.lev, nd.lev)
}
\arguments{
\item{e.lev}{existing data levels}

\item{nd.lev}{new data levels}
}
\description{
Examine the levels of relevant factors in new data and make
them conform to the levels of the relevant factors on which the model is
based. This means placing unknown levels after known levels for factors that
have unknown levels in the new data, or adding missing factor levels and
sorting the levels for factors that have levels that are not present in the
new data
}
\author{
Dan Putler
}
