% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs.R
\name{mrsDataObj}
\alias{mrsDataObj}
\title{Create MRS data object}
\usage{
mrsDataObj(con_string, table, fields, loc_tmp, rem_tmp, rem_R)
}
\arguments{
\item{con_string}{connection string}

\item{table}{table}

\item{fields}{fields}

\item{loc_tmp}{loc_tmp}

\item{rem_tmp}{rem_tmp}

\item{rem_R}{rem_R}
}
\description{
Take the variables to be used in an analysis along with information about the
table and the platform (SQL Server or Teradata) and then creates a data source
object with alpha sorted factor levels.
}
\author{
Dan Putler
}
