% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs.R
\name{mrsGetXYLevels}
\alias{mrsGetXYLevels}
\title{Get MRS levels.}
\usage{
mrsGetXYLevels(data, fields)
}
\arguments{
\item{data}{data}

\item{fields}{fields}
}
\description{
Determine the levels of any predictor variables that are factors based on
information in an object created by mrsDataObj
}
\author{
Dan Putler
}
