% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs.R
\name{mrsLevels}
\alias{mrsLevels}
\title{Get MRS levels}
\usage{
mrsLevels(data, loc_tmp, rem_tmp = NULL, rem_R = NULL)
}
\arguments{
\item{data}{data}

\item{loc_tmp}{loc_tmp}

\item{rem_tmp}{rem_tmp}

\item{rem_R}{rem_R}
}
\description{
Get the names of potential factors in a table
Note 1: Applying names() to the returned mrsLevels objects, the names of the
factor variable can be determined.
Note 2: The set of factor variables altered in subsequent steps can be
addressed by using an approach such as
mrs_levels[names(mrs_levels) %in% desired_factor_names]
}
\author{
Dan Putler
}

