% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs.R
\name{mrsLevels2ColInfo}
\alias{mrsLevels2ColInfo}
\title{Extract column info from MRS levels.}
\usage{
mrsLevels2ColInfo(levels)
}
\arguments{
\item{levels}{x levels of MRS data object}
}
\description{
Take the xlevels element of a predictive model object and convert into a
structrue that is correct as the \code{colInfo} argument to a MRS
in-platform data object
}
\author{
Dan Putler
}
