% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs.R
\name{mrsReorderedLevels}
\alias{mrsReorderedLevels}
\title{Reorder MRS Labels}
\usage{
mrsReorderedLevels(lev_obj)
}
\arguments{
\item{lev_obj}{mrs levels object}
}
\description{
Take the original factor levels in a mrsLevels object and reorder them
using an alpha sort on the level labels.
}
\author{
Dan Putler
}

