% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs.R
\name{mrsReorderFactors}
\alias{mrsReorderFactors}
\title{Create the appropriate information for the \code{colInfo} arguments of
\code{RxSqlServerData} and \code{RxTeradata} to get factor
levels in alpha order}
\usage{
mrsReorderFactors(data, loc_tmp, rem_tmp, rem_R)
}
\arguments{
\item{data}{data}

\item{loc_tmp}{loc_tmp}

\item{rem_tmp}{rem_tmp}

\item{rem_R}{rem_R}
}
\description{
Create the appropriate information for the \code{colInfo} arguments of
\code{RxSqlServerData} and \code{RxTeradata} to get factor
levels in alpha order
}
\author{
Dan Putler
}

