% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs.R
\name{mrsTemp}
\alias{mrsTemp}
\title{Create a temp directory for MRS related work}
\usage{
mrsTemp()
}
\description{
\code{mrsTemp} determines if an appropriate temporary directory for MRS related
work is present on the user's machine, creates one if not, and returns the
path to the directory
}
\author{
Dan Putler
}
