% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreutils.R
\name{noNullLevels}
\alias{noNullLevels}
\title{Remove factors with null levels from a list}
\usage{
noNullLevels(ll)
}
\arguments{
\item{ll}{list of factor levels.}
}
\description{
Remove factors with null levels from a list
}

