% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisiontreeutils.R
\name{outputDTResultsAlteryx}
\alias{outputDTResultsAlteryx}
\title{output results to Alteryx}
\usage{
outputDTResultsAlteryx(results, config)
}
\arguments{
\item{results}{list of results to output}

\item{config}{list of config options}
}
\description{
output results to Alteryx
}

