% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestedTests.R
\name{performNestedTest}
\alias{performNestedTest}
\title{Perform a nested test on linear models}
\usage{
performNestedTest(inputs)
}
\arguments{
\item{inputs}{list of objects consisting of 1 data frame and 2 model objects.}
}
\description{
This function performs a nested test for linear and generalized linear models
}
\examples{
mod1 <- lm(mpg ~ ., data = mtcars)
mod2 <- lm(mpg ~ wt, data = mtcars)
inputs <- list(
  list(Name = 'mod1', Object = mod1),
  list(Name = 'mod2', Object = mod2),
  mtcars
)
performNestedTest(inputs)
}
\author{
Dan Putler, Ramnath Vaidyanathan
}
