% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonutils.R
\name{pipeDelim}
\alias{pipeDelim}
\title{Convert row into a pipe delimited string}
\usage{
pipeDelim(capture.vec)
}
\arguments{
\item{capture.vec}{row vector to convert into pipe delimited string}
}
\description{
The function pipeDelim takes a table row that comes from capture.output and
turns it into a pipe delimited string for processing in Alteryx.
}
\author{
Dan Putler
}
