% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonutils.R
\name{plotMeans}
\alias{plotMeans}
\title{Plot means by group(s)}
\usage{
plotMeans(response, factor1, factor2, error.bars = c("se", "sd", "conf.int",
  "none"), level = 0.95, xlab = deparse(substitute(factor1)),
  ylab = paste("mean of", deparse(substitute(response))),
  legend.lab = deparse(substitute(factor2)), main = "Plot of Means",
  pch = 1:n.levs.2, lty = 1:n.levs.2, col = palette(), ...)
}
\arguments{
\item{response}{the response variable}

\item{factor1}{the first grouping variable}

\item{factor2}{the second grouping variable}

\item{error.bars}{the type of error bars to plot}

\item{level}{the confidence level to use when plotting conf.int}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{legend.lab}{legend label}

\item{main}{main title of the plot}

\item{pch}{plotting character i.e. symbol to use}

\item{lty}{line type}

\item{col}{color palette to use}

\item{...}{additional arguments. unused currently.}
}
\description{
This function is taken from John Fox's plotMeans function from the Rcmdr package
}

