% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisiontreeutils.R
\name{processDT}
\alias{processDT}
\title{Process DT model}
\usage{
processDT(inputs, config)
}
\arguments{
\item{config}{list of configuration options}

\item{data}{list of datastream objects}
}
\value{
list of results or results
}
\description{
Process DT model
}

