% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularizationutils.R
\name{processElasticNet}
\alias{processElasticNet}
\title{Process Elastic Net Inputs}
\usage{
processElasticNet(inputs, config)
}
\arguments{
\item{inputs}{input data streams to the tool}

\item{config}{configuration passed to the tool}
}
\description{
This function takes `inputs` and `config` and returns the model object
along with other elements essential to create the reports and plots
}
