% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearregutils.R
\name{processLinearOSR}
\alias{processLinearOSR}
\alias{processLinearXDF}
\title{Process Linear Regression Inputs}
\usage{
processLinearOSR(inputs, config)

processLinearXDF(inputs, config)
}
\arguments{
\item{inputs}{input data streams to the tool}

\item{config}{configuration passed to the tool}
}
\description{
These two functions take `inputs` and `config` and return the model object
along with other elements essential to create the reports and plots
}
