% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearregutils.R
\name{processLinearOSR}
\alias{processLinearOSR}
\title{#### Process Inputs}
\usage{
processLinearOSR(inputs, config)
}
\description{
These two function take `inputs` and `config` and return the model object
along with other elements essential to create the reports and plots
}

