% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticregutils.R
\name{processLogisticOSR}
\alias{processLogisticOSR}
\title{Process Inputs}
\usage{
processLogisticOSR(inputs, config)
}
\description{
These two function take `inputs` and `config` and return the model object
along with other elements essential to create the reports and plots
}

