% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonutils.R
\name{pStars}
\alias{pStars}
\title{Convert p-values to text.}
\usage{
pStars(p.val)
}
\arguments{
\item{p.val}{vector of pvalues.}
}
\description{
The function pStars takes a vector of p-values and (1) converts them to text
and formats the "top-end" to be consistent with standard R reporting of
p-values and (2) creates a text column of "stars" to signify significance
levels.
}
\author{
Dan Putler
}
