% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{readModelObjects}
\alias{readModelObjects}
\title{Read an Alteryx stream consisting of a table of model objects}
\usage{
readModelObjects(name, default)
}
\arguments{
\item{name}{name of the Alteryx input to read.}

\item{default}{default to be used when run outside of an Alteryx workflow.}
}
\description{
Read an Alteryx stream consisting of a table of model objects
}

