% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rSquared}
\alias{rSquared}
\title{Tries very hard to compute a sensible R squared, else returns NaN.}
\usage{
rSquared(numeric_vector_1, numeric_vector_2)
}
\arguments{
\item{numeric_vector_1}{one numeric vector}

\item{numeric_vector_2}{another numeric vector of the same length}
}
\description{
Tries very hard to compute a sensible R squared, else returns NaN.
}
