% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtree_cross_validation.R
\name{runCrossValidationDTree}
\alias{runCrossValidationDTree}
\title{Wrapper function for performing cross-validation on DTree models.}
\usage{
runCrossValidationDTree(inputs, config)
}
\arguments{
\item{inputs}{list containing a Dtree model and the data used to generate it}

\item{config}{list of config options}
}
\value{
list with components measuring model fit obtained via CV
}
\description{
Wrapper function for performing cross-validation on DTree models.
}
