% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinRegCV.R
\name{runCrossValidationLinReg}
\alias{runCrossValidationLinReg}
\title{process for Linear Regression with Cross Validation
returns named vector of six model-performance metrics
r_squared, adj_r_squared, avg_mae, avg_mape, avg_mse, avg_rmse}
\usage{
runCrossValidationLinReg(inputs, config)
}
\arguments{
\item{inputs}{list of inputs}

\item{config}{list of config elements}
}
\description{
process for Linear Regression with Cross Validation
returns named vector of six model-performance metrics
r_squared, adj_r_squared, avg_mae, avg_mape, avg_mse, avg_rmse
}

