% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogRegCV.R
\name{runCrossValidationLogReg}
\alias{runCrossValidationLogReg}
\title{Run CV and output results for logistic CV model}
\usage{
runCrossValidationLogReg(inputs, config)
}
\arguments{
\item{inputs}{inputs list with data and models}

\item{config}{list of configuration options}
}
\value{
list of results for outputting
}
\description{
Run CV and output results for logistic CV model
}
