% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModels.R
\name{runLogisticRegression}
\alias{runLogisticRegression}
\title{Modify the link so that it can be passed on to R.}
\usage{
runLogisticRegression(inputs, config)
}
\description{
Modify the link so that it can be passed on to R.
}

