% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreModel.R
\name{scoreModel2}
\alias{scoreModel.rxLinMod}
\alias{scoreModel2}
\alias{scoreModel2.default}
\alias{scoreModel2.lm}
\alias{scoreModel2.rxDTree}
\alias{scoreModel2.rxGlm}
\alias{scoreModel2.rxLogit}
\title{Score function}
\usage{
scoreModel2(mod.obj, new.data, score.field = "Score", ...)

\method{scoreModel2}{default}(mod.obj, new.data, score.field = "Score",
  os.value = NULL, os.pct = NULL, ...)

\method{scoreModel2}{lm}(mod.obj, new.data, score.field = "Score",
  pred.int = FALSE, int.vals = NULL, log.y = FALSE, ...)

\method{scoreModel2}{rxLogit}(mod.obj, new.data, score.field = "Score",
  os.value = NULL, os.pct = NULL, ...)

\method{scoreModel2}{rxGlm}(mod.obj, new.data, score.field = "Score", ...)

scoreModel.rxLinMod(mod.obj, new.data, score.field = "Score",
  pred.int = FALSE, int.vals = NULL, log.y = FALSE, ...)

\method{scoreModel2}{rxDTree}(mod.obj, new.data, score.field, os.value = NULL,
  os.pct = NULL, ...)
}
\arguments{
\item{mod.obj}{model object}

\item{new.data}{new data to score}

\item{score.field}{name given to the score field}

\item{...}{additional arguments}

\item{os.value}{oversampling value}

\item{os.pct}{oversampling percent}

\item{pred.int}{whether to generate prediction intervals}

\item{int.vals}{interval values}

\item{log.y}{whether to report y on the log scale}
}
\description{
Score function
}
\author{
Ramnath Vaidyanathan, Dan Putler
}

