% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{serializeObject}
\alias{serializeObject}
\title{Serialize an arbitrary R object into an ASCII string}
\usage{
serializeObject(the.object)
}
\arguments{
\item{the.object}{object to serialize}
}
\description{
This function converts an arbitrary R object and serializes it into an ASCII
string.  Spatial objects become Well Known Text
}
\author{
Rob Bryan
}
