% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVutils.R
\name{setPositiveClass}
\alias{setPositiveClass}
\title{Set the postive class for two-class classification.}
\usage{
setPositiveClass(tar_lev, order)
}
\arguments{
\item{tar_lev}{a vector of string with the levels of target variable}

\item{order}{string - choices are "alpha" and "common".
For "alpha", use first by alphabetical order.
For "common", use less common class.}
}
\value{
string - name of positive class
}
\description{
Set the postive class for two-class classification.
}

