% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonutils.R
\name{standardize}
\alias{standardize}
\title{Standardize a matrix}
\usage{
standardize(x, std.vec1, std.vec2)
}
\arguments{
\item{x}{matrix to standardize}

\item{std.vec1}{vector to subtract from columns}

\item{std.vec2}{vector to divide the columns}
}
\description{
The function takes a matrix and two standardization vectors and standardizes
the variables based on subtracting the first standardization vector from the
appropriate columns and then dividing by the second standarization parameter.
}
\author{
Dan Putler
}

