% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonutils.R
\name{trim.blanks}
\alias{trim.blanks}
\title{Trim Blanks Function}
\usage{
trim.blanks(text)
}
\arguments{
\item{text}{input text that may contain blanks}
}
\description{
This function trims blanks from text input
}
