% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonutils.R
\name{unitScale}
\alias{unitScale}
\title{Unit Scale Function}
\usage{
unitScale(x)
}
\arguments{
\item{x}{a numeric matrix object}
}
\description{
The unitScale function standardizes variables so that their values fall in the
inclusive unit interval. The argument (x) must be an object that is, or can be
coerced to be, a numeric matrix. The function returns a matrix of values whose
columns have been standardized to have values that fall in the inclusive unit
interval.
}
\author{
Dan Putler
}
