% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestedTests.R
\name{unserializeInputs}
\alias{unserializeInputs}
\title{Unserialize a list of inputs some of which might contain serialized model objects}
\usage{
unserializeInputs(inputs)
}
\arguments{
\item{inputs}{list of inputs}
}
\description{
Unserialize a list of inputs some of which might contain serialized model objects
}
\author{
Dan Putler, Ramnath Vaidyanathan
}

