% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{unserializeObject}
\alias{unserializeObject}
\title{Unserialize a string produced by serializeObject}
\usage{
unserializeObject(the.string)
}
\arguments{
\item{the.string}{string to unserialize}
}
\description{
The function unserializeObject takes a string produced by serializeObject
and unserializes it
}
\author{
Rob Bryan
}

