% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonutils.R
\name{validName}
\alias{validName}
\title{Create a valid name for an R object}
\usage{
validName(name)
}
\arguments{
\item{name}{name of the object}
}
\description{
The function validName takes an R object name as its argument, then and (1)
determines if it is a valid R object name and (2) if not, make it one.
}
\author{
Dan Putler
}
