% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defensiveutils.R
\name{verifyClass}
\alias{verifyClass}
\title{Check if positive class is present in vector
Will return same value if in vector, value from data if level is in vector
with different case / trim, and an empty string if no similar value is present}
\usage{
verifyClass(v, lvl)
}
\arguments{
\item{v}{vector to check for presence of class}

\item{lvl}{level of class to check for presence of}
}
\value{
positive class
}
\description{
Check if positive class is present in vector
Will return same value if in vector, value from data if level is in vector
with different case / trim, and an empty string if no similar value is present
}
