% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonutils.R
\name{wrapTable}
\alias{wrapTable}
\title{Wrap a data frame object into a table.}
\usage{
wrapTable(dframe, width = 6, first.rnames = TRUE, round.level = 5e-07)
}
\arguments{
\item{dframe}{A data frame (or an object that can be coerced to a data frame)}

\item{width}{The maximum number of columns to include in each table. If the
first column contains rows names, this column is included in the width of
the table.}

\item{first.rnames}{A logical value which indicates if the first column of
the table contains row names that should be included in all the new tables
as the first column of each table.}

\item{round.level}{level of rounding.}
}
\value{
The function returns a character vector, each element of which
  contains a row of the table, including the column names, with the cell
  values in each row delimited by pipes.
}
\description{
The function wrapTable takes a data frame object, or an object that can be
coerced to be a data frame (such as a matrix) and breaks it into two or more
data frames, each with a fixed number of columns, but which collectively have
all the columns of the original table. The rows of the corresponding tables
are then converted to pipe delimited strings, and the full collection of rows
is returned as a character vector.
}
\author{
Dan Putler
}

