% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localization.R
\name{XMSG}
\alias{XMSG}
\title{localization.R
translates input string and embeds any bind-variable values in result}
\usage{
XMSG(in.targetString_sc, in.firstBindVariable_sc = NULL,
  in.secondBindVariable_sc = NULL, in.thirdBindVariable_sc = NULL,
  in.fourthBindVariable_sc = NULL, in.fifthBindVariable_sc = NULL,
  in.sixthBindVariable_sc = NULL)
}
\arguments{
\item{in.targetString_sc}{the input string to translate}

\item{in.firstBindVariable_sc}{the first bind variable}

\item{in.secondBindVariable_sc}{the second bind variable}

\item{in.thirdBindVariable_sc}{the third bind variable}

\item{in.fourthBindVariable_sc}{the fourth bind variable}

\item{in.fifthBindVariable_sc}{the fifth bind variable}

\item{in.sixthBindVariable_sc}{the sixth bind variable}
}
\value{
translated string with embedded bind-variable values (if any)
}
\description{
localization.R
translates input string and embeds any bind-variable values in result
}

