% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localization.R
\name{xmsg_and_pipe_delimit}
\alias{xmsg_and_pipe_delimit}
\title{translates strings in input vector and separates them with pipes}
\usage{
xmsg_and_pipe_delimit(in.term_vc)
}
\arguments{
\item{in.term_vc}{vector of terms to XMSG and pipe-delimit}
}
\value{
single translated, pipe-delimited string
}
\description{
translates strings in input vector and separates them with pipes
}

