% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adj_r_squared}
\alias{adj_r_squared}
\title{Translates an R squared into an adjusted R squared, returning NaN
if the input is Nan.}
\usage{
adj_r_squared(r_squared, n, p, intercept_degrees_freedom)
}
\arguments{
\item{r_squared}{R squared}

\item{n}{number of records (rows)}

\item{p}{number of independent variables (predictors)}

\item{intercept_degrees_freedom}{one if model includes intercept,
else zero}
}
\description{
Translates an R squared into an adjusted R squared, returning NaN
if the input is Nan.
}

