% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisiontreeutils.R
\name{adjustCP}
\alias{adjustCP}
\title{Adjusts config based on results if config was initially "Auto"}
\usage{
adjustCP(model, config)
}
\arguments{
\item{model}{model object}

\item{config}{list of config options}
}
\value{
model obj after adjusting complexity parameter
}
\description{
Adjusts config based on results if config was initially "Auto"
}

