% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtree_cross_validation.R
\name{areIdentical}
\alias{areIdentical}
\title{Cross-validation for Decision Tree}
\usage{
areIdentical(v1, v2)
}
\arguments{
\item{v1}{a vector}

\item{v2}{a vector}

\item{config}{list of config options}

\item{the.data}{incoming data}
}
\value{
list of results or results

Boolean indicating whether the two vectors are equal
}
\description{
Cross-validation for Decision Tree
}

