% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisiontreeutils.R
\name{checkValidConfig}
\alias{checkValidConfig}
\title{General Method for Validating config
Error checking pre-model
Returns a vector of Booleans indicating warnings and errors}
\usage{
checkValidConfig(config, the.data, names)
}
\arguments{
\item{config}{list of config options}

\item{the.data}{incoming data}

\item{names}{list of x, y, w names for data}
}
\value{
messages_vb vector of booleans
}
\description{
General Method for Validating config
Error checking pre-model
Returns a vector of Booleans indicating warnings and errors
}

