% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisiontreeutils.R
\name{convertDTParamsToArgs}
\alias{convertDTParamsToArgs}
\title{name mapping from parameters to functions}
\usage{
convertDTParamsToArgs(f_string, params)
}
\arguments{
\item{f_string}{string of function}

\item{params}{list of decision tree params}
}
\value{
list with named parameters for f_string
}
\description{
name mapping from parameters to functions
}

