% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisiontreeutils.R
\name{convertDTParamsToArgs}
\alias{convertDTParamsToArgs}
\title{Map parameter names for OSR and XDF}
\usage{
convertDTParamsToArgs(params, f_string)
}
\arguments{
\item{params}{list of decision tree params}

\item{f_string}{string of function}
}
\value{
list with named parameters for f_string
}
\description{
Map parameter names for OSR and XDF
}

