% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisiontreeutils.R
\name{createDashboardDT}
\alias{createDashboardDT}
\title{Create Interactive Dashboard}
\usage{
createDashboardDT(model, is_XDF)
}
\arguments{
\item{model}{model object}

\item{is_XDF}{boolean of whether model is XDF}
}
\description{
Create Interactive Dashboard
}

