% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisiontreeutils.R
\name{createDTParams}
\alias{createDTParams}
\title{Creation of components for model object evaluation}
\usage{
createDTParams(config, names, xdf_properties)
}
\arguments{
\item{config}{list of config options}

\item{xdf_properties}{list of xdf details (is_XDF and xdf_path elements)}

\item{data}{list of datastream inputs}
}
\value{
list with components needed to create model
}
\description{
Creation of components for model object evaluation
}

