% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtree_cross_validation.R
\name{createFolds}
\alias{createFolds}
\title{Create the list of cross-validation folds}
\usage{
createFolds(data, config)
}
\arguments{
\item{data}{the data.frame used to create the models}

\item{config}{a list of configuration information}
}
\value{
list of record ID's. Each element is the record ID's of the folds for a given trial.
}
\description{
Create the list of cross-validation folds
}

