% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisiontreeutils.R
\name{createReportDT}
\alias{createReportDT}
\title{Get data for static report (grp|out pipes)}
\usage{
createReportDT(model, config, names, is_XDF)
}
\arguments{
\item{model}{model object}

\item{config}{list of config options}

\item{names}{names of variables (x, y and w)}

\item{is_XDF}{boolean of whether model is XDF}
}
\value{
dataframe of piped results
}
\description{
Get data for static report (grp|out pipes)
}

