% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisiontreeutils.R
\name{createTreePlotDT.C5.0}
\alias{createTreePlotDT.C5.0}
\title{Create Tree Plot for C5.0 model}
\usage{
\method{createTreePlotDT}{C5.0}(model, config)
}
\arguments{
\item{model}{model object}

\item{config}{configuration object}
}
\value{
graphs
}
\description{
Create Tree Plot for C5.0 model
}

