% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisiontreeutils.R
\name{createTreePlotDT}
\alias{createTreePlotDT}
\title{Create Tree Plot}
\usage{
createTreePlotDT(model, config, inputs)
}
\arguments{
\item{model}{model object}

\item{config}{configuration object}

\item{inputs}{input list}
}
\value{
graphs
}
\description{
Create Tree Plot
}
