% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisiontreeutils.R
\name{createTreePlotDT.rpart}
\alias{createTreePlotDT.rpart}
\title{Create Tree Plot for rpart model}
\usage{
\method{createTreePlotDT}{rpart}(model, config)
}
\arguments{
\item{model}{model object}

\item{config}{configuration object}
}
\value{
graphs
}
\description{
Create Tree Plot for rpart model
}

