% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisiontreeutils.R
\name{createTreePlotDT.rxDTree}
\alias{createTreePlotDT.rxDTree}
\title{Create Tree Plot for rxDTree model}
\usage{
\method{createTreePlotDT}{rxDTree}(model, config, inputs)
}
\arguments{
\item{model}{model object}

\item{config}{configuration object}

\item{inputs}{input list}
}
\value{
graphs
}
\description{
Create Tree Plot for rxDTree model
}

