% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createmodelutils.R
\name{getNamesFromOrdered}
\alias{getNamesFromOrdered}
\title{Get names list for x, y, and weights}
\usage{
getNamesFromOrdered(names, use_weights)
}
\arguments{
\item{names}{vector of names for data}

\item{use_weights}{boolean for whether weights are in data}
}
\value{
list of x, y, and weights with respective names
}
\description{
Get names list for x, y, and weights
}

