% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisiontreeutils.R
\name{getOutputsDT}
\alias{getOutputsDT}
\title{get results in form of output list}
\usage{
getOutputsDT(config, model, is_XDF)
}
\arguments{
\item{config}{list of config options}

\item{model}{model object}

\item{is_XDF}{boolean of whether model is XDF}
}
\description{
get results in form of output list
}

