% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xdfutils.R
\name{getXdfLevels}
\alias{getXdfLevels}
\title{extract levels from xdf data}
\usage{
getXdfLevels(formula, xdf)
}
\arguments{
\item{xdf}{path to xdf file}

\item{form}{R formula}
}
\value{
names of levels - vector
}
\description{
extract levels from xdf data
}

