% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonutils.R
\name{graphWHR}
\alias{graphWHR}
\title{GraphWHR Function}
\usage{
graphWHR(inches = c("True", "False"), in.w, in.h, cm.w = NULL,
  cm.h = NULL, resolution = c("1x", "2x", "3x"), print.high = FALSE)
}
\description{
The function graphWH takes information about the desired graph size (in
inches or centemeters), the desired output resolution of 1x, 2x, 3x from a
base of 96 dpi, and and (in the case of 3x) whether high (576 dpi) resolution
should be used to determine the appropriate width, height, and res values to
provide the AlteryxGraph device.
}
\author{
Dan Putler
}

