% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactiveLMReport.R
\name{interactive_lm_report}
\alias{interactive_lm_report}
\title{linear-model interactive-report function
This function generates the interactive report for the Linear Model tool.}
\usage{
interactive_lm_report(config, inputs, model)
}
\arguments{
\item{config}{config}

\item{inputs}{inputs}

\item{model}{model}
}
\description{
linear-model interactive-report function
This function generates the interactive report for the Linear Model tool.
}
\author{
Todd Morley, Dylan Blanchard
}

