% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactiveLMReport.R
\name{interactive_lm_report}
\alias{interactive_lm_report}
\title{Dashboard generation for linear regression models}
\usage{
interactive_lm_report(config, data, model)
}
\arguments{
\item{config}{config}

\item{data}{data (breaking the usual mold a little)}

\item{model}{model}
}
\value{
dashboard object for rendering
}
\description{
Dashboard generation for linear regression models
}

