% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisiontreeutils.R
\name{processDT}
\alias{processDT}
\title{Process DT model}
\usage{
processDT(inputs, config)
}
\arguments{
\item{inputs}{input data streams to the tool}

\item{config}{configuration passed to the tool}
}
\value{
list of results or results
}
\description{
Process DT model
}

