% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisiontreeutils.R
\name{processDT}
\alias{processDT}
\title{process for converting to results list from config and data}
\usage{
processDT(config, data)
}
\arguments{
\item{config}{list of configuration options}

\item{data}{list of datastream objects}
}
\value{
list of results or results
}
\description{
process for converting to results list from config and data
}

