% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinRegCV.R
\name{runCrossValidationLinReg}
\alias{runCrossValidationLinReg}
\title{process for Linear Regression with Cross Validation}
\usage{
runCrossValidationLinReg(inputs, config)
}
\arguments{
\item{inputs}{list of inputs}

\item{config}{list of config elements}
}
\description{
process for Linear Regression with Cross Validation
}

